增加了@ResultMapper
在moviesmapper和tag里增加了@ResultMapper，并通过了测试。
