+++ src/main/java/com/dbcourse/curriculum_design/mapper/TagsMapper.java
@@ -2,10 +2,7 @@ package com.dbcourse.curriculum_design.mapper;
 
 import com.dbcourse.curriculum_design.model.Tags;
 import com.dbcourse.curriculum_design.model.TagsExample;
-import org.apache.ibatis.annotations.Mapper;
-import org.apache.ibatis.annotations.Param;
-import org.apache.ibatis.annotations.ResultMap;
-import org.apache.ibatis.annotations.Select;
+import org.apache.ibatis.annotations.*;
 
 import java.util.List;
 
@@ -38,5 +35,6 @@ public interface TagsMapper {
     @Select("select top (#{num, jdbcType=INTEGER})" +
             " n_id, c_content " +
             " from Tags ")
+    @ResultMap("BaseResultMap")
     List<Tags> getTopNumTags(int num);
 }
