+++ src/main/java/com/dbcourse/curriculum_design/mapper/UsersAndShortCommentsMapper.java
@@ -25,7 +25,7 @@ public interface UsersAndShortCommentsMapper {
     int batchInsert(@Param("list") List<UsersAndShortComments> list);
 
     @Select("select * from UsersAndShortComments order by ShortCommentsId DESC " +
-            "offset ((#{pageIndex,jdbcType=INTEGER}-1)*#{pageSize,jdbcType=INTEGER}) rows " +
+            "offset ((#{pageIndex,jdbcType=INTEGER}-1) * #{pageSize,jdbcType=INTEGER}) rows " +
             "fetch next #{pageSize,jdbcType=INTEGER} rows only")
     List<UsersAndShortComments> getShortCommentsByPage(int pageIndex, int pageSize);
 }
