+++ src/main/java/com/dbcourse/curriculum_design/service/UsersAndLongCommentsService.java
@@ -24,7 +24,7 @@ public interface UsersAndLongCommentsService {
 
     int batchInsert(List<UsersAndLongComments> list);
 
-}
+
 
     /**
      *
