+++ src/main/java/com/dbcourse/curriculum_design/service/impl/UsersAndLongCommentsServiceImpl.java
@@ -55,7 +55,7 @@ public class UsersAndLongCommentsServiceImpl implements UsersAndLongCommentsServ
         return usersAndLongCommentsMapper.batchInsert(list);
     }
 
-}
+
 
     @Override
     public List<UsersAndLongComments> selectByMovieId(Integer MovieId) {
